Split runtime host client boundaries
